refactor: 위치 검증 로직 개선 적용
- 채팅방 생성 시:
  - 사용자 위치 정보를 Request Body(`ChatRoomRequest`)에 포함하여 요청하도록 변경함.
  - `ChatRoomService.setCreateChatRoom`에서 PostGIS 기반 반경 검증(`isUserWithinFestivalRadius`)을 수행하며, 반경 밖일 경우 `BusinessException`을 발생시켜 **생성 자체를 거부**함.
  - 반경 검증 통과 후 위치 토큰(`redisTokenService.setToken`)을 발급/갱신함.

- 채팅방 입장 시 (WebSocket /join):
  - **입장은 위치 인증 여부와 관계없이 자유롭게 허용**하도록 로직을 변경함.
  - `ChatLocationService.setLocationTokenOnJoin`을 호출하여 사용자 위치를 체크함.
  - **반경 내일 경우에만** 위치 인증 토큰을 발급/갱신하고, 반경 밖일 경우 **예외 없이 토큰 발급/갱신을 건너뜀** (`return`).
  - 실제 메시지 전송(`send`) 시에만 위치 인증 토큰의 유효성(`validateChatRoomEntry`)을 검사하여 행위를 제한함.
- 테스트 코드 반영

Ref: #46

+++ common-lib/src/main/java/com/grm3355/zonie/commonlib/domain/chatroom/entity/ChatRoom.java
@@ -59,7 +59,7 @@ public class ChatRoom extends BaseTimeEntity {
 	private String coverImageUrl;
 
 	@Column(name = "max_participants", nullable = false)
-	private int maxParticipants;
+	private Long maxParticipants;
 
 	@Column(name = "radius", nullable = false)
 	private double radius;
